% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_mem.R
\name{create_mem_output}
\alias{create_mem_output}
\title{create MEM season plots}
\usage{
create_mem_output(conf, date)
}
\arguments{
\item{conf}{A mem model configuration object}

\item{date}{extract date}
}
\description{
create MEM season plots
}
