% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_email_obs.R
\name{EmailExternalGenerateTable}
\alias{EmailExternalGenerateTable}
\title{Generates the outbreak table for the external email}
\usage{
EmailExternalGenerateTable(results, xtag, xemail)
}
\arguments{
\item{results}{a}

\item{xtag}{a}

\item{xemail}{a}
}
\description{
Generates the outbreak table for the external email
}
