% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_alert_pdf.R
\docType{data}
\name{sykdomspuls}
\alias{sykdomspuls}
\title{normomo}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sykdomspuls
}
\description{
normomo
}
\keyword{datasets}
