% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_mem.R
\docType{data}
\name{sykdomspuls_mem}
\alias{sykdomspuls_mem}
\title{Mem -outputs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sykdomspuls_mem
}
\description{
Mem -outputs
}
\keyword{datasets}
