% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_email_obs.R
\docType{data}
\name{sykdomspuls_obs}
\alias{sykdomspuls_obs}
\title{normomo}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sykdomspuls_obs
}
\description{
normomo
}
\keyword{datasets}
