% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspuls_email_obs.R
\name{sykdomspuls_obs_email_external}
\alias{sykdomspuls_obs_email_external}
\title{Sends an external email warning about alters}
\usage{
sykdomspuls_obs_email_external()
}
\description{
Sends an external email warning about alters
}
