% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspulspdf.R
\docType{data}
\name{sykdomspulspdf}
\alias{sykdomspulspdf}
\title{Mem -outputs}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
sykdomspulspdf
}
\description{
Mem -outputs
}
\keyword{datasets}
