% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_base.R
\name{UIBase}
\alias{UIBase}
\title{Mem -outputs}
\description{
Mem -outputs

Mem -outputs
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-run_with_catch}{\code{UIBase$run_with_catch()}}
\item \href{#method-run_all}{\code{UIBase$run_all()}}
\item \href{#method-clone}{\code{UIBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_with_catch"></a>}}
\if{latex}{\out{\hypertarget{method-run_with_catch}{}}}
\subsection{Method \code{run_with_catch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UIBase$run_with_catch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\if{latex}{\out{\hypertarget{method-run_all}{}}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UIBase$run_all()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UIBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
