% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_download_weather.R
\docType{data}
\name{weather_download}
\alias{weather_download}
\title{amort}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
weather_download
}
\description{
amort
}
\keyword{datasets}
