% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sykdomspulspdf.R
\name{sykdomspulspdf}
\alias{sykdomspulspdf}
\title{Mem -outputs}
\description{
Mem -outputs

Mem -outputs
}
\section{Super class}{
\code{\link[ui:UIBase]{ui::UIBase}} -> \code{sykdomspulspdf}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-run_all}{\code{sykdomspulspdf$run_all()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="ui" data-topic="UIBase" data-id="run_with_catch">}\href{../../ui/html/UIBase.html#method-run_with_catch}{\code{ui::UIBase$run_with_catch()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\subsection{Method \code{run_all()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sykdomspulspdf$run_all()}\if{html}{\out{</div>}}
}

}
}
